Add update invoice logic and views Add download invoice (docx) Add command to migrate from postgres to sqlite

+++ core/management/commands/clear_migrations.py
@@ -11,6 +11,7 @@ class Command(BaseCommand):
         migrations_folder = os.path.join("**", "migrations", "*.py")
         # glob.glob returns a list of paths matching a pathname pattern
         for file in glob.glob(migrations_folder, recursive=True):
-            if not file.endswith("__init__.py"):
+            # exclude core app migrations (custom migrations added manually) and __init__.py files
+            if not file.startswith("core") and not file.endswith("__init__.py"):
                 os.remove(file)
                 self.stdout.write(self.style.SUCCESS("Successfully deleted file %s" % file))
